+++ native/packages/react-native-bpk-component-card/src/withDivider-test.ios.js
@@ -16,8 +16,8 @@
  * limitations under the License.
  */
 
-import BpkCard from './src/BpkCard';
-import withDivider from './src/withDivider';
+import commonTests from './withDivider-test.common';
 
-export { withDivider };
-export default BpkCard;
+describe('iOS', () => {
+  commonTests();
+});